+++ src/main/java/com/adaweng/shoppingcart/entity/Coupon.java
@@ -1,5 +1,6 @@
 package com.adaweng.shoppingcart.entity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Transient;
@@ -11,8 +12,8 @@ public class Coupon {
 	private String code;
 	private Date startDate;
 	private Date endDate;
-	private Double priceReached;
-	private Double priceReduced;	
+	private BigDecimal priceReached;
+	private BigDecimal priceReduced;	
 	private User user;
 	@Transient
 	private String userId;
@@ -47,16 +48,16 @@ public class Coupon {
 	public void setEndDate(Date endDate) {
 		this.endDate = endDate;
 	}
-	public Double getPriceReached() {
+	public BigDecimal getPriceReached() {
 		return priceReached;
 	}
-	public void setPriceReached(Double priceReached) {
+	public void setPriceReached(BigDecimal priceReached) {
 		this.priceReached = priceReached;
 	}
-	public Double getPriceReduced() {
+	public BigDecimal getPriceReduced() {
 		return priceReduced;
 	}
-	public void setPriceReduced(Double priceReduced) {
+	public void setPriceReduced(BigDecimal priceReduced) {
 		this.priceReduced = priceReduced;
 	}
 	public User getUser() {
