+++ src/main/java/com/adaweng/shoppingcart/entity/Product.java
@@ -1,10 +1,12 @@
 package com.adaweng.shoppingcart.entity;
 
+import java.math.BigDecimal;
+
 public class Product {
 	private Long id;
 	private String name;
 	private String type;
-	private Double unitPrice;
+	private BigDecimal unitPrice;
 	private String description;
 	private Long stockAmount;
 	private String status;
@@ -33,10 +35,10 @@ public class Product {
 	public void setDescription(String description) {
 		this.description = description;
 	}
-	public Double getUnitPrice() {
+	public BigDecimal getUnitPrice() {
 		return unitPrice;
 	}
-	public void setUnitPrice(Double unitPrice) {
+	public void setUnitPrice(BigDecimal unitPrice) {
 		this.unitPrice = unitPrice;
 	}
 	public Long getStockAmount() {
